+++ lm/@stk_lm_/disp.m
@@ -1,8 +1,8 @@
-% STK_COVMAT_NOISE [STK internal]
+% DISP [overload base function]
 
 % Copyright Notice
 %
-%    Copyright (C) 2019 CentraleSupelec
+%    Copyright (C) 2021 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -26,10 +26,8 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function K = stk_covmat_noise (model, varargin)
+function disp (lm)
 
-K = stk_covmat_noise (model.prior, varargin{:});
+fprintf ('<%s>\n', stk_sprintf_sizetype (lm));
 
 end % function
-
-%#ok<*INUSD,*STOUT>