+++ lm/@stk_lm_/stk_isnoisy.m
@@ -1,8 +1,8 @@
-% STK_COVMAT_NOISE [STK internal]
+% STK_ISNOISY [overload STK funcion]
 
 % Copyright Notice
 %
-%    Copyright (C) 2019 CentraleSupelec
+%    Copyright (C) 2021 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -26,10 +26,8 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function K = stk_covmat_noise (model, varargin)
+function b = stk_isnoisy (model)  %#ok<INUSD>
 
-K = stk_covmat_noise (model.prior, varargin{:});
+b = false;
 
 end % function
-
-%#ok<*INUSD,*STOUT>