+++ lm/@stk_lm_sincos/feval.m
@@ -1,8 +1,8 @@
-% STK_COVMAT_NOISE [STK internal]
+% FEVAL [overload base function]
 
 % Copyright Notice
 %
-%    Copyright (C) 2019 CentraleSupelec
+%    Copyright (C) 2021 CentraleSupelec
 %
 %    Author:  Julien Bect  <julien.bect@centralesupelec.fr>
 
@@ -26,10 +26,12 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function K = stk_covmat_noise (model, varargin)
+function z = feval (lm, x)
 
-K = stk_covmat_noise (model.prior, varargin{:});
+x = double (x);  assert (iscolumn (x));
+
+u = lm.angular_frequency * x;
+
+z = [sin(u), cos(u)];
 
 end % function
-
-%#ok<*INUSD,*STOUT>