+++ model/prior_struct/stk_covmat_noise.m
@@ -34,8 +34,19 @@
 %    You should  have received a copy  of the GNU  General Public License
 %    along with STK.  If not, see <http://www.gnu.org/licenses/>.
 
-function [K, P1, P2] = stk_covmat_noise (model, x1, x2, diff, pairwise)
+function K = stk_covmat_noise (model, varargin)
 
+if stk_isnoisy (model)
+    K = stk_covmat_noise_ (model, varargin{:});
+else
+    K = stk_nullcov ([], varargin{:});
+end
+
+end % function
+
+
+function K = stk_covmat_noise_ (model, x1, x2, diff, pairwise)
+    
 % Number of evaluations points
 n1 = size (x1, 1);
 if (nargin > 2) && (~ isempty (x2))
@@ -53,7 +64,7 @@ if nargin < 4,  diff = -1;  end
 pairwise = (nargin > 4) && pairwise;
 assert ((n1 == n2) || (~ pairwise));
 
-if autocov && (diff ~= 0) && (stk_isnoisy (model))
+if autocov && (diff ~= 0)
     
     if ~ isnumeric (model.lognoisevariance)
         
@@ -123,7 +134,9 @@ else  % Return a null matrix
     % a) autocov is false: we are actually computing a *cross*-covariance matrix
     % b) diff = 0: derivative with respect to a parameter that does not modify
     %    the covariance matrix of the noise
-    % c) we are in the noiseless case.
+    
+    % TEMP: Does b) really happen?  I bet not.
+    assert (diff ~= 0);
     
     if pairwise
         K = zeros (n1, 1);
@@ -131,18 +144,6 @@ else  % Return a null matrix
         K = zeros (n1, n2);
     end
     
-end
-
-
-%% Compute matrices for the linear part
-
-% No linear part for the 'noise' output: return empty matrices if required
-if nargout > 1
-    P1 = zeros (n1, 0);
-    
-    if nargout > 2
-        P2 = zeros (n2, 0);
-    end
 end
 
 end % function
@@ -157,43 +158,36 @@ end % function
 %! x1 = stk_sampling_randunif (n1, d);
 %! x2 = stk_sampling_randunif (n2, d);
 
-%!error [KK, PP] = stk_covmat_noise ();
-%!error [KK, PP] = stk_covmat_noise (model);
+%!error KK = stk_covmat_noise ();
+%!error KK = stk_covmat_noise (model);
 
-%!test  [Ka, Pa] = stk_covmat_noise (model, x1);                           % (1)
-%!test  [K1, P1] = stk_covmat_noise (model, x1, []);
-%!test  [K2, P2] = stk_covmat_noise (model, x1, [], -1);
-%!test  [K3, P3] = stk_covmat_noise (model, x1, [], -1, false);
+%!test  Ka = stk_covmat_noise (model, x1);                           % (1)
+%!test  K1 = stk_covmat_noise (model, x1, []);
+%!test  K2 = stk_covmat_noise (model, x1, [], -1);
+%!test  K3 = stk_covmat_noise (model, x1, [], -1, false);
 %!assert (isequal (size (Ka), [n1 n1]));
-%!assert (isequal (size (Pa), [n1 0]));
-%!assert (isequal (P1, Pa) && (isequal (K1, Ka)))
-%!assert (isequal (P2, Pa) && (isequal (K2, Ka)))
-%!assert (isequal (P3, Pa) && (isequal (K3, Ka)))
+%!assert (isequal (K1, Ka))
+%!assert (isequal (K2, Ka))
+%!assert (isequal (K3, Ka))
 
-%!test  [Kb, Pb] = stk_covmat_noise (model, x1, x1);                       % (2)
-%!test  [K1, P1] = stk_covmat_noise (model, x1, x1, -1);
-%!test  [K2, P2] = stk_covmat_noise (model, x1, x1, -1, false);
+%!test  Kb = stk_covmat_noise (model, x1, x1);                       % (2)
+%!test  K1 = stk_covmat_noise (model, x1, x1, -1);
+%!test  K2 = stk_covmat_noise (model, x1, x1, -1, false);
 %!assert (isequal (size (Kb), [n1 n1]));
-%!assert (isequal (size (Pb), [n1 0]));
-%!assert (isequal (P1, Pb) && (isequal (K1, Kb)))
-%!assert (isequal (P2, Pb) && (isequal (K2, Kb)))
+%!assert (isequal (K1, Kb))
+%!assert (isequal (K2, Kb))
 
-%!test  [Kc, Pc] = stk_covmat_noise (model, x1, x2);                       % (3)
-%!test  [K1, P1] = stk_covmat_noise (model, x1, x2, -1);
-%!test  [K2, P2] = stk_covmat_noise (model, x1, x2, -1, false);
+%!test  Kc = stk_covmat_noise (model, x1, x2);                       % (3)
+%!test  K1 = stk_covmat_noise (model, x1, x2, -1);
+%!test  K2 = stk_covmat_noise (model, x1, x2, -1, false);
 %!assert (isequal (size (Kc), [n1 n2]));
-%!assert (isequal (size (Pc), [n1 0]));
-%!assert (isequal (P1, Pc) && (isequal (K1, Kc)))
-%!assert (isequal (P2, Pc) && (isequal (K2, Kc)))
+%!assert (isequal (K1, Kc))
+%!assert (isequal (K2, Kc))
 
 % In the noiseless case, (1) and (2) should give the same results
 %!assert (isequal (Kb, Ka));
 
 % In the noisy case, however...
-%!test  [Ka, Pa] = stk_covmat_noise (model2, x1);                         % (1')
-%!test  [Kb, Pb] = stk_covmat_noise (model2, x1, x1);                     % (2')
+%!test  Ka = stk_covmat_noise (model2, x1);                         % (1')
+%!test  Kb = stk_covmat_noise (model2, x1, x1);                     % (2')
 %!error assert (isequal (Kb, Ka));
-
-% The second output depends on x1 only => should be the same for (1)--(3)
-%!assert (isequal (Pa, Pb));
-%!assert (isequal (Pa, Pc));