+++ NEWS.md
@@ -13,7 +13,25 @@
 -----
 
 
-# Changes in version 2.7.0
+# Changes in version 2.7.0  (OR SHOULD THIS BE 3.0.0 ???)
+
+## Parameter estimation criteria
+
+* `stk_param_relik.m`: the old syntax for computing the gradient of
+   the restricted likelihood function:
+
+        [C, COVPARAM_DIFF, LNV_DIFF] = stk_param_relik (...)
+
+  is deprecated.  Instead, use:
+
+        [C, C_GRAD] = crit (...)
+
+  to obtain the full gradient with respect to all optimizable
+  parameters of the model.  (Note the change of behaviour for the
+  syntax with two output arguments, if the model has optimizable
+  parameters in the noise model, for instance.)
+
+* `stk_param_proflik`: NEW
 
 ## Required Matlab version